:recycle: decoBtn utils 함수 분리
